rework command
